pyproject.toml
[tool.poetry]
name = "place-remember"
version = "0.1.0"
description = ""
authors = ["Maxim <max2004@zel@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
django-allauth = "^0.54.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
